// src/app/models/ReportModel.ts
export class BPCReportGRR extends CommonClass {
    ID: Number;
    Client: String;
    Company: String;
    Type: String;
    PatnerID: String;
    Material: String;
    GRGIDoc: string;
    GRIDate: Date;
    Item: string;
    Description: string;
    MaterialText: String;
    GRGIQty: number;
    OrderQty: Number;
    ReceivedQty: Number;
    RejectedPPM: Number;
    ReworkQty: Number;
    Status: String;
}